Fixed numerous warnings. Split tinyxml into its' own library so that it would not warn.

+++ src/com_wiz.cpp
@@ -361,7 +361,7 @@ BOOL CMDSstatus::Execute(const std::string &verb, Player* mobile,std::vector<std
     st << "userspace cpu usage: " << TimevalToString(&(usage.ru_utime)) << "." << std::endl;
     st << "Kernel cpu usage: " << TimevalToString(&(usage.ru_stime)) << "." << std::endl;
 //memory:
-    st << "Max rss: " << usage.ru_maxrss << "KB (" << (float)(usage.ru_maxrss/1024.0F) << "mb)." << std::endl;
+    st << "Max rss: " << usage.ru_maxrss << "KB (" << (usage.ru_maxrss/1024.0F) << "mb)." << std::endl;
     st << Repete("-", 80) << std::endl;
 //timing information
     updates = world->GetUpdates();
